browse_use/pyproject.toml
[project]
name = "browse-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.2.5",
    "langchain-google-genai>=2.1.2",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]
